pyproject.toml
[project]
name = "ballot-proofer-functions"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "azure-ai-documentintelligence>=1.0.2",
    "azure-core>=1.34.0",
    "azure-functions>=1.23.0",
    "azure-identity>=1.22.0",
    "azure-storage-blob>=12.25.1",
    "dotenv>=0.9.9",
    "openai>=1.77.0",
]

[dependency-groups]
dev = [
    "ruff>=0.11.8",
]
